+++ extlibs/qwt-6.0.1/src/qwt_plot_spectrocurve.cpp
@@ -222,7 +222,7 @@ void QwtPlotSpectroCurve::drawSeries( QPainter *painter,
         return;
 
     if ( to < 0 )
-        to = dataSize() - 1;
+        to = (int)dataSize() - 1;
 
     if ( from < 0 )
         from = 0;
